pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arb-engine"
version = "0.1.0"
description = "CEX/PerpDEX Arbitrage Bot with Real-time Monitoring"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web Framework & UI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "nicegui>=1.4.0",

    # Exchange APIs
    "ccxt>=4.2.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",

    # Data & Storage
    "duckdb>=0.10.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pyarrow>=15.0.0",

    # Configuration & Validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",

    # Logging & Monitoring
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",

    # Utilities
    "httpx>=0.26.0",
    "tenacity>=8.2.3",  # Retry logic
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
]

[project.scripts]
arb-engine = "arb_engine.app.main:main"
arb-collector = "arb_engine.scripts.run_collector:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/arb_engine/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
